requirements.txt
pytest
requests
selenium
Appium-Python-Client